pyproject.toml
[tool.poetry]
name = "extadel-romanian"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
langchain = "^0.0.320"
openai = "^0.28.1"
tiktoken = "^0.5.1"
numpy = "^1.26.1"
pandas = "^2.1.1"
python-dotenv = "^1.0.0"
urllib3 = "^2.0.7"
streamlit = "^1.27.2"
streamlit-agraph = "^0.0.45"
pydantic-settings = "^2.0.3"
wordcloud = "^1.9.2"
matplotlib = "^3.8.0"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"
pep8 = "^1.7.1"
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
